requirements.txt
appdirs==1.4.4
asgiref==3.3.0
astroid==2.4.2
attrs==20.3.0
autopep8==1.5.4
bcrypt==3.2.0
cached-property==1.5.2
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
colorama==0.4.4
cryptography==3.2.1
defusedxml==0.6.0
Django==3.1.3
django-cors-headers==3.5.0
django-filter==2.4.0
django-oauth-toolkit==1.3.3
djangorestframework==3.12.2
idna==2.10
isodate==0.6.0
isort==5.6.4
keyring==21.5.0
lazy-object-proxy==1.4.3
lxml==4.6.2
mccabe==0.6.1
oauth2-provider==0.0
oauthlib==3.1.0
paramiko==2.7.2
pycodestyle==2.6.0
pycparser==2.20
pycryptodome==3.9.9
pylint==2.6.0
pylint-django==2.3.0
pylint-plugin-utils==0.6
PyNaCl==1.4.0
pysftp==0.2.9
pytz==2020.4
pywin32-ctypes==0.2.0
requests==2.24.0
requests-file==1.5.1
requests-toolbelt==0.9.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.25.11
wrapt==1.12.1
xmltodict==0.12.0
zeep==4.0.0
